Refactored layout and login as temporary solution til zustand is implemented
