+++ app/(auth)/signup/page.tsx
@@ -1,9 +1,8 @@
 'use client'
-import { useState } from 'react'
 import { useForm, SubmitHandler } from 'react-hook-form'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
-import signIn from '../actions/signIn'
+import signUp from '../../actions/signUp'
 import {
   Form,
   FormControl,
@@ -14,7 +13,7 @@ import {
 } from '@/components/ui/form'
 import { Input } from '@/components/ui/input'
 import { Button } from '@/components/ui/button'
-import Link from 'next/link'
+import { useState } from 'react'
 import {
   Card,
   CardContent,
@@ -23,36 +22,51 @@ import {
   CardHeader,
   CardTitle,
 } from '@/components/ui/card'
+import Link from 'next/link'
 import { Loader2 } from 'lucide-react'
+import { formatPhoneNumber } from '@/lib/utils'
 
 // Interface for form values
-interface LoginFormValues {
+interface SignupFormValues {
   email: string
   password: string
+  displayName: string
+  phone: string
 }
 
 // Zod schema for form validation
-const loginSchema = z.object({
+const signupSchema = z.object({
   email: z.string().email({ message: 'Invalid email format' }),
   password: z
     .string()
-    .min(6, { message: 'Password must be at least 6 characters long' }),
+    .min(6, { message: 'Password must be at least 6 characters' }),
+  displayName: z.string().min(1, { message: 'Display name is required' }),
+  phone: z
+    .string()
+    .min(10, { message: 'Phone number must be at least 10 digits' }),
 })
 
-export default function Login() {
-  const [signInError, setSignInError] = useState('')
+export default function Signup() {
+  const [signUpError, setSignUpError] = useState('')
   const [isLoading, setIsLoading] = useState(false)
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<SignupFormValues>()
 
-  const form = useForm<LoginFormValues>({
-    resolver: zodResolver(loginSchema),
+  const form = useForm<SignupFormValues>({
+    resolver: zodResolver(signupSchema),
   })
 
-  const onSubmit: SubmitHandler<LoginFormValues> = async (data) => {
+  const onSubmit: SubmitHandler<SignupFormValues> = async (data) => {
     setIsLoading(true)
     try {
-      await signIn(data)
+      await signUp(data)
+      if (typeof window !== 'undefined') {
+      }
     } catch (error) {
-      setSignInError('Sign-in failed. Please check your credentials.')
+      setSignUpError('Sign-up failed. Please try again.')
     } finally {
       setIsLoading(false)
     }
@@ -62,7 +76,7 @@ export default function Login() {
     <div className="flex min-h-screen items-center justify-center">
       <Card className="flex flex-col items-center justify-center py-2">
         <CardHeader>
-          <CardTitle>Login</CardTitle>
+          <CardTitle>Sign Up</CardTitle>
           <CardDescription>Please enter your credentials</CardDescription>
         </CardHeader>
         <CardContent>
@@ -78,11 +92,9 @@ export default function Login() {
                   <FormItem>
                     <FormLabel>Email</FormLabel>
                     <FormControl>
-                      <Input placeholder="Email" {...field} />
+                      <Input type="email" placeholder="Email" {...field} />
                     </FormControl>
-                    <FormMessage>
-                      {form.formState.errors.email?.message}
-                    </FormMessage>
+                    <FormMessage />
                   </FormItem>
                 )}
               />
@@ -99,13 +111,44 @@ export default function Login() {
                         {...field}
                       />
                     </FormControl>
-                    <FormMessage>
-                      {form.formState.errors.password?.message}
-                    </FormMessage>
+                    <FormMessage />
                   </FormItem>
                 )}
               />
-              {signInError && <p className="text-red-500">{signInError}</p>}
+              <FormField
+                control={form.control}
+                name="displayName"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel>Display Name</FormLabel>
+                    <FormControl>
+                      <Input placeholder="Display Name" {...field} />
+                    </FormControl>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+              <FormField
+                control={form.control}
+                name="phone"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel>Phone Number</FormLabel>
+                    <FormControl>
+                      <Input
+                        type="tel"
+                        placeholder="Phone Number"
+                        {...register('phone', { required: true })}
+                        onChange={(e) => {
+                          e.target.value = formatPhoneNumber(e.target.value)
+                        }}
+                      />
+                    </FormControl>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+              {signUpError && <p className="text-red-500">{signUpError}</p>}
               {isLoading ? (
                 <Button disabled>
                   <Loader2 className="mr-2 h-4 w-4 animate-spin" />
@@ -116,18 +159,21 @@ export default function Login() {
                   type="submit"
                   className="mb-2 rounded-md bg-green-700 px-4 py-2 text-white hover:bg-green-600"
                 >
-                  Log In
+                  Sign Up
                 </Button>
               )}
             </form>
           </Form>
         </CardContent>
-        <CardFooter>
+        <CardFooter className="flex flex-col">
           <Link
-            href="/signup"
+            href="/login"
             className="text-center text-sm text-foreground/60"
           >
-            Don't have an account? Sign up
+            Already have an account? Login
+          </Link>
+          <Link href="/" className="text-center text-sm text-foreground/60">
+            Go Home
           </Link>
         </CardFooter>
       </Card>
